Fix error in seeding issues
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/121544
introduced a regression in seeding
issues. `Gitlab::Usage::MetricDefinition.paths` is an array of paths
that need to be globbed, not a an array of
filenames. `metric_definitions_changed?` failed because it attempted
to retrieve the modification time of a glob instead of a filename.  To
fix the error, process the path as a glob first and find the last
change from all files.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1874