Add RSS feed on topic page
This feature adds a RSS feed for topics, so that we can follow topics we
like and be notified when new projects are being released.

In order to avoid spamming the same projects multiple times, the
projects are sorted by id, thus creation date. So the feed is about
discovering new projects, and not, for example, getting news on updates
for existing projects (for that, an user may follow the RSS feed on the
project activity page).

Changelog: added