Revert docs update: all API error messages are already user facing
At least in the case of the REST API, all API error messages are already
user facing. We do not expose internal server exception messages.

So we don't need to add prefixes to API responses, which will be
confusing to REST API and GraphQL API consumers

Because we remove the section in api_styleguide.md, we also have to
remove the corresponding section in fe_guide/style/javascript.md.