Migrate resource_link_events to ghost users
When a user is destroyed, all rows that reference the user
should either accept NULL for user or have the reference
migrated to ghost user.

`resource_link_events` table had a non-nullable reference to
`users` table but did not have the required migration step
to ghost users in MigrateRecordsToGhostUserService
and it has blocked the process of user deletion.

Changelog: security
EE: true