Make review-apps mandatory in MRs again
review-apps are currently stable, and the main
failures we had in the last two months were due
to the fact that review-apps were not mandatory
in MRs that had failures.

Those failiures got into the main branch and broke
all the review-apps.

This commit ensures that we block failed review-apps
deployments before an MR is allowed to be merged.