Move wait_for_gitlab_to_respond to QA::Page::Base
After the last change the method couldn't be used as a class method,
which is expected of all methods where module_function is used.

This moves the method to QA::Page::Base where it will still be available
to all page objects. At the moment it's only used in
QA::Page::Main::Login.