Add service for merge when approved and pipelines succeeds strategy
This change introduces a service for a new auto merge strategy.
It is similar to "merge when pipeline succeeds", but can also be enabled
before approvals are given. When this strategy is enabled and pipeline
finishes before the approvals are given, the MR gets merged with the
final required approval given.