Add "auditors" parameter to users API
Allow all auditors to be queried like all admins can be from
the administrator access to the users API.

Ensure that the query is hidden behind the check for the add-on to
a license.

Add a post-deployment database migration which adds an index for the
auditor column with the ID of the user.

Changelog: added
EE: true
feature: addition