Update template question
`How do I create a template?` question is not clear because it doesn't
specify what kind of template is meant. So technically the correct
answer to this should be something like `What kind of template do you
mean?`. But because the original purpose of this question was rather
test that documentation-related question is answered, it makes sense to
update the question to be more specific.