Move Secret Detection SHA logic to Finding
This is a more appropriate place to determine the correct SHA. The
Finding model is shared by the Vulnerability model as well as the
::Vulnerabilities::FindingPresenter class.

This fixes an issue where the incorrect blob path was being shown for
Secret Detection findings within the MR Security Widget. This is because
the Widget uses an API driven by Projects::MergeRequestsController which
makes use of ::Vulnerabilities::FindingPresenter to render each Secret
Detection finding JSON object.

Hopefully this change ensures that the SHA is correctly computed
regardless of who is sourcing the finding.