Generate Slack app manifest for self-managed
The GitLab for Slack app that is distibuted through the Slack App
Directory can only work with GitLab.com. This is because the Slack app
configuration requires specific hard-coded URLs, so these are all set
to gitlab.com.

This change allows self-managed customers to create a copy of the GitLab
for Slack app to install into their Slack workspace.

The Slack app they create is functionally identical to the official
public one.

The change is behind a feature flag.

https://gitlab.com/gitlab-org/gitlab/-/issues/358872