Return deploy key title when humanizing access levels
When we humanize a protected ref access level with a deploy key we used
to return "Deploy Key", however, if we look at how users and groups are
humanized we return their names which identifies the user or group
rather than the role.

As a deploy key is not a role but rather a thing we should return the
name so users have more context.

Changelog: changed