[docs] Update `cron` docs to call out special `fugit` syntax
I was writing a `cron` job. I saw this example, however the docs say
the examples are from the `man` page or from wikipedia

This confused me, as both suggest that `%` is invalid and creates a
newline

Took a few minutes for me to realize it is part of the `fugit` parser
extensions.

I'd like to make that more clear by calling it out in-line somewhere
closer to the example