Adds table ml_model and entity Ml::Model
A Ml::Model has:
- `name`, which will be used to reference Packages::Package by
  package_name (follow up: a table packages_ml_model_metadata will be
added to add an association for a package of type ml_model to a model).
`name` must be a valid ml_model package_name. Name is unique per project
and acts as an iid of sorts.
- `project`, project which the model belongs to
- `default_experiment`, an Ml::Experiment with the same name as the
  model. When creating the model, and an experiment with the same name
exists, that is set as the default_experiment. Otherwise a new
experiment is created. An experiment that is the default experiment
cannot be destroyed.

Changelog: added