Add cost factored storage size to graphql ProjectStatistics response
The cost factored storage size is the total storage size of the project
with any cost factor for forks applied.
A fork may have a cost factor applied based on its visibility and the
root namespace subscription.
For a project that is not a fork, the cost factored storage size will
always equal the total storage size.