Remove diffs app pre-renderer
This pre-renderer adds additional overhead for the sake of always
being able to scroll to the exact same place.
Instead of this overhead happening on load of any diffs file
this has been moved to our sync component.
The sync component scrolls to the right index and then waits
for the size to be defined before scrolling again.
Doing this means we reduce the amount of time spent on the main
thread when loading the page.

Remove the `renderIt` queue from the diffs app.
The `renderIt` queue was used before we had the virtual scroller,
with the virtual scroller we are always setting `renderIt` to
`true` so the additional time spent by creating callbacks
and calling Vuex commits isn't needed.

https://gitlab.com/gitlab-org/gitlab/-/issues/338359