Clean up orphaned `Vulnerabilities::Remediation` records
The process for removing `remediations` in a pipeline leaves these
`remediations` in the database without any `findings` associated with
them. As a result, these remediations become orphaned in the database.

It was discovered that ~40% of all the `remediations` in the production
database are actually orphaned.

This commit adds a simple Cron worker that deletes any remediations
where the `findings` are blank