Make removable checks less expensive
The previous query loads the entire solo-owned groups to memory and
checked the presence using .present? method. This causes memory bloat
as loading 10k+ groups to memory at the same time will consume
Gigabytes of memory for a single request.

This commit fixes this behaviour by delegating this check to
the database.

Changelog: performance