Fix enabling shared runners on group
if override is allowed

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/124728
we started to update `allow_descendants_override_disabled_shared_runners`
when enabling shared runners.

It turns out that it breaks the old validation, which was
validating the fact of changing the field instead the end state.

This MR fixes the validation and adds some test.