Include users from shares in autocomplete
When querying users for the group-level autocomplete, we include users
that are direct members of child projects. This MR extends this so that
it includes members via projects shared to groups.

We also currently include direct members of subgroups. This MR extends
this so it includes members via group shares.