Remove cache param from MergeToRefService
This feature is never used. Past attempts at enabling it have been
reverted. The first one caused an incident with merge trains, and the
second caused a bug with stale merge request diffs.

The implementation was also incorrect, because the cache key did not
contain all the parameters sent to the gitaly RPC, and some of these can
vary between calls (for example, the user).

Since the code is both unused and incorrect, and the caching was not
important enough to revisit this a third time over the last two years,
we are removing it entirely.