Change build queuing metrics FF to ops type
This flag has been enabled for a long time on GitLab.com, and the
metrics are considered production-ready, but the they are undocumented
and stuck behind a `:development` flag, making it largely inaccessible
to self-managed customers.

To address those issues, we are making this flag to an `:ops` type flag
and adding documentation. The flag remains disabled by default to avoid
adding unexpected load on customers' Prometheus instances.

See https://gitlab.com/gitlab-org/gitlab/-/issues/350888

Changelog: changed