Add AI-gateway architecture blueprint
This moves the AI-gateway architecture that was discussed in
https://gitlab.com/gitlab-com/gl-infra/readiness/-/merge_requests/166
to the main repository for visibility. It started in the readiness
document as a simple braindump, but then the discussion happened there
so we decided to continue it there.

But now that the first iteration has been accepted, we think it might
be more accessible for all stakeholders as an architecture
blueprint. I'll make sure to crosslink the old location to this one if
this has been accepted.

It doesn't follow the exact template of a blueprint, but does contain
all of the information. I changed some of the titles to match and
added a link to the issue where we discussed alternatives. The rest is
from the original MR.

For https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/161