Add test for deleted or moved docs
The changes in this commit add a test for deleted or renamed
documentation files. This test will check if a deleted or renamed
documentation file is still present in the GitLab Docs global
navigation. If the file is still present, the test will fail and the
merge request will not be able to be merged. This will help to ensure
that documentation files are not deleted or renamed without a redirect
being added.