Expose deploy key in push access level graphql type
Branch rules requires us to display when a deploy key is assigned to the
push access levels.

Prior to this change we didn't return deploy key's
as a type.

In EE we expose `user` and `group` as their own types so clients can
differentiate between role access levels. Used as inspiration to expose
`deploy_key`s.

Adds a new policy rule to `DeployKeyPolicy` to enable `read_deploy_key`
if the deploy key is public or the user has access to a project where
the key is assigned.

Changelog: added