Ensure components are ingested in a consistent order
When generating SBoMs, Trivy sorts components by the `bomRef`
attribute, which can either be the package URL of the dependency, or a
UUID. When it is a UUID, it results in components being in a random
order. When this happens, it causes deadlocks when ingesting SBoMs, as
the inserts acquire locks on random rows and this can sometimes result
in circular locking. This change ensures that components are always
ordered consistently prior to ingestion, which will allow the locks to
be acquired and released in order.

Changelog: fixed
EE: true