doc/api: Clarify group storage moves API use case
The group storage moves API is used to migrate group wiki repositories
from one storage to another. Projects contained within a group are not
migrated as part of this operation. In the documentation, this is not
immediately clear to the reader and it would be reasonable to assume
that moving a group to a new storage would also move all associated
project repositories.

Clarify the group storage moves API documentation introduction to
explicitly explain that a group's project repositories are not moved
through the use of this API.