Avoid calling tree with ref_type when using sha
The ref_type parameter is needed when the tree needs to be serialized
with context about whether the ref is a branch or a tag.  It's also
used for creating a fully qualified ref to find the correct commit.
This is not neccessary and actually incorrect when the ref is the sha
for a commit.

Contributes to: https://gitlab.com/gitlab-org/gitlab/-/issues/20526