Fix canApprove GraphQL field returning null
The `Approvable#eligible_for_unapproval_by?` can return `nil`
when `user` is not present. The `canApprove` GraphQL field is
required and not nullable though. As a result, it returns a
GraphQL error as a response to GraphQL query.

To fix that, we check if `user.present?` instead of only `user`.
Also updated the spec to check for `true` and `false` value to
ensure that it doesn't return a `nil` value.

Changelog: fixed