Use DS report with latest GLAD identifier links
Since continuous scanning will always scan against the
latest version of advisories, we should not compare
against a report that uses a fixed version of the gem-
nasium db a.k.a. the GitLab Advisory Database. Thus,
the report will now reference the `master` ranch in
all of the advisory identifier urls.