Fix group path update validations
Previously, when a group path was updated, we would check if any NPM
packages existed under that group.
This was wrong as not all NPM packages can impact the group path update.
In reality, only NPM packages that follow the naming convention need to
be checked.

The validation logic was thus updated to only consider:
* Root groups
* Namespaced NPM packages

This change is behind a feature flag.

Changelog: fixed