Fix file variable expansion on downstream pipeline
This change is behind feature flag ci_prevent_file_var_expansion_downstream_pipeline.

Previously, variables passed into downstream pipelines that
contain reference on file variables have the file variable content expanded into a
new environment variable.

When this feature flag is turned on, the file variables are not expanded.
Instead, the file variables that are needed for expansion are forwarded
as file variables into the downstream pipelines.
This allows the downstream pipeline jobs to access the file variables
as files in the filesystem.