Refactor all Zoekt HTTP calls to Zoekt::Client
This extracts a new class `Gitlab::Search::Zoekt::Client` which becomes
the central place for making HTTP requests to Zoekt. Prior to this
change the indexing and search calls were split into different places.
This is probably a better code organization but mainly this is a small
refactoring to make it easier to implement
https://gitlab.com/gitlab-org/gitlab/-/issues/389749 which adds basic
auth credentials to our HTTP requests. We could combine that into this
MR but it seemed preferable to review this independently.