Creates fake names for Ml::Candidate
If a candidate is not defined, creates a new random composed of
'animal-animal-animal-number'. This will improve the UI and allow users
to search better through candidates. Existing candidates that don't have
a name will continue to not have a name.

There are 141 animal names in
https://github.com/ffaker/ffaker/blob/main/lib/ffaker/data/animal/common_names,
so that means there are 28B (141*141*141*10000) possible combinations
per project. Names can be duplicated.

Changelog: changed