Do not store chat messages by default
We no longer want to store and show AI messages on the chat, if not
explicitly enabled by the feature. It is now only enabled for the `chat`
AI action.
We do this by setting `skip_cache = true` by default.

It also fixes a bug where the `skip_cache` was not passed along properly
to the GraphqlSubscriptionResponseService.

Changelog: fixed
EE: true