Use URL helpers for security policy project readme links
Ensures that documentation links go to the correct version of the docs
for the current version of GitLab. For example, a self-hosted GitLab
running version 16.2 will go to the 16.2 docs hosted on the instance,
instead of docs.gitlab.com which may have a newer version of the docs
like 16.3.

Changelog: changed
EE: true