Fix navigation error when clicking active pipeline tab
When the page first loaded the activeTabName value was null as we are
trying to fetch the default value from validPipelineTabNames which
doesn't include pipelineTabName as the pipeline tab is the root URL.

To fix this I've added a check to see if the current path is '' and then
returned pipelineTabName.

I've also modified navigateTo to check if the clicked tab is already
active.

Users likely weren't noticing this error so no changelog entry.