Requires Ml::ModelVersion to be valid SemVer
If we enforce semver, we will be able to add features for the model
registry, like automatic update of models. We can only do this now,
since this feature hasn't been released yet, and we call rollback this
later, but supporting semver once this feature is out will be much
harder.

Changelog: changed