Extract regex modules into discrete files, add MR module
Extracts Packages and BulkImports modules into their own files, and adds
a new module `Gitlab::Regex::MergeRequests` for MR-related regexes. The
goal of this shuffling is to reduce complexity of `lib/gitlab/regex.rb`
and make it a bit less of a "junk drawer" class, while establishing a
pattern for additional cleanups, and providing a home for future regexes to land in.