Add basic auth to Zoekt client
As part of https://gitlab.com/gitlab-org/gitlab/-/issues/389749 we want
HTTP requests from `Zoekt -> GitLab` to be authenticated. This MR adds
the ability to configure a `zoekt.username_file` and
`zoekt.password_file` (that default to `.gitlab_zoekt_username` and
`.gitlab_zoekt_password` respectively) in your `config/gitlab.yml` .
If present these files will be used to populate basic auth credentials
in the HTTP requests to Zoekt. We likely won't use this in GDK because
basic authentication relies on an nginx proxy in front of Zoekt (this is
configured in the helm chart in
https://gitlab.com/gitlab-org/cloud-native/charts/gitlab-zoekt/-/merge_requests/14).
The indexer and webserver we run in GDK don't actually support basic
auth but I've verified locally that these changes do indeed send the
credentials.

EE: true
Changelog: added