gitlab/git: Provide regexps to match object SHAs
We have a bunch of callsites where we want to match an object SHA that
is optionally abbreviated. This requires us to hardcode both the minimum
and maximum object SHA length, which makes it hard to change them at a
global scale.

Introduce a new `SHA_PATTERN` constant and adapt the code to use it.
This eases the transation towards recognizing SHA256 object SHAs.