Refactor of ContentSecurityPolicy::ConfigLoader
The complexity of the `ConfigLoader::default_directives` method was
quite high.

When the complexity is high it becomes harder to read and understand
what a method is doing.

With this refactor the `default_directives` method is less cluttered.
Conditions for directives are now moved inside their respective methods.
Methods are reorganized in order they are called.
Instance method moved below class methods.