Allow frontend to access LFS file storage URL
When loading the blob page the blob controller inserts the connect_src
values for the files that are displayed on the requested page.

When VueJS navigates to the blob page content security policy from the
tree page will still be active, meaning the download request for the
file will be blocked by the content security policy as the connect_src
directive doesn't include the storage url.

Now we check if lfs storage is enabled and proxy download is turned off
during initialization. If they both are true we add the storage url to
the global content security policy.

Changelog: fixed