Remove feature flag
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/124229 it was
simpler to handle the logic as the one liner rather than
conditionally defining the constant to include or not include 'ref'.

Now that the feature flag is being removed and the behavior has been
rolled out to production, I think it is simpler to remove the one liner
and remove 'ref' from `ALLOWED_SORT_VALUES`.

The logic in the affected method `build_sort_params` already does the
desired behavior if we remove 'ref' from the method.

This commit also:
- removes the feature flag yaml.
- updates the spec for feature flag removal.

Changelog: fixed