Send archived to the indexer if the Project is archived
We want to hide by default the blobs of the archived projects, so we
need to index the archived value of the project in the blobs index.
To do this we will send archived flag if the project is archived.

Changelog: other
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/126680
EE: true