Add an application setting to enable public package registries
This adds the package_registry_allow_anyone_to_pull_option boolean field
to Application Settings.

When enabled, the toggle "Allow anyone to access the Package Registry"
works as before. It is visible for Private and Internal projects, and
hidden for Public projects.

When disabled, the toggle "Allow anyone to access the Package Registry"
cannot be enabled. The toggle is not shown for non-Public projects.

Changelog: added