Fix restrict_user_defined_variables restriction on pipeline schedules
When implementing restrict_user_defined_variables, we missed adding it
to pipeline schedules. In this commit, we are also adding it to
its create and update service.

If restrict_user_defined_variables is enabled;
- On create; when the user is a developer and is creating a pipeline
schedule with variables, then the action is forbidden.
- On update; when the user is a developer and is updating a pipeline
schedule with variables, then the action is forbidden.

Changelog: fixed