Add includeSourceCode to `ExplainVulnerability` cache key
While working on the frontend of the "Explain This Vulnerability"
feature, it was discovered that the cached AI response does not
differentiate between the `includeSourceCode: true` and
`includeSourceCode: false` versions of a request.

It will return whichever one it has in its cache, even if the user
changes the checkbox selection.

We need to address this because it would be alarming for a user if
they unchecked the checkbox, yet they receive a response that was
clearly passed the source code.

This is because the cache key was not factoring in any request options

This change includes the `include_source_code` option as part of the
cache key

Rather than splatting the entire `options` hash into the key, I have
opted to be explicit in the `options` I check for

Changelog: fixed
EE: true