Fix order-dependent ActionCable spec failure
If another spec uses ActionCable before the ActionCable initializer
spec runs, then the current PubSub Redis connection will already be
configured with a memoized identifier. The spec attempts to test a
Redis configuration with a `nil` identifier to support GCP
Memorystore, but for that to work we need to reset the Redis
connection and cause the new config to be used.

This commit fixes the order-dependent test by clearing out the
ActionCable server status before and after the test via
`ActionCable.server.restart`.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/418937