Ensure is_primary is set on TrainingProviderUpdate
The Graphql mutation 'SecurityTrainingUpdate' has the `isPrimary` field
marked as not required, however the corresponding field is required in
the model that gets created by the mutation - `Security::Training`.

This can cause a 500 error from `Pg::NotNullViolation` on the column in
these scenarios:

- The `isPrimary` argument is not provided in the mutation
- The `isPrimary` argument is explicitly set to `null` in the mutation

Both of these are valid from the GQL standpoint when `required` is set
to `false` on an argument.

This changes ensures that `isPrimary` defaults to `false` if not
provided or explicitly set to `null`.

The specs have also been moved to request specs as per our Graphql
guidelines and additional specs added to cover this bug.