Add Places group to default search view
This copies the previously-named "persistent links" from the context
switcher dropdown to the default view of the global search modal.

The recent issues and merge requests links are now only shown when in
a context which has them. This avoids sending issue/merge request
dashboard links to non-logged in users, who would just be redirected to
the login page anyway.

This also preserves/extends a feature added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/114285, which
hides issue-related links from users in projects with the feature
disabled.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/409522.

Changelog: added