Move unified rules alert to help text of approvers
As we have migrated to providing multiple approver rules configuration
via the UI for a few milestones, we can now move the loud alert warning
users to use the API for unified rules to the subtle form group.

All the links still exist, as does the documentation, we are just
reducing visual noise.

Changelog: changed
EE: true