Extract MergeService merge strategies into a class
Currently `MergeRequests::MergeService` is bloated, handling multiple
responsibilities and relying on circular monkey patching inheritance.
This makes it hard to expand and add new features.

This MR splits out the responsibility for the git merge operation into a
class `MergeRequests::MergeStrategies::FromSourceBranch`. This class is
responsible for preparing the source for merge (including an optional
squash), and performing the git merge operation (including determining
whether a fast-forward or merge commit is needed). This leaves
`MergeRequests::MergeService` exclusively responsible for the reusable
application logic associated with a merge, such as updating database
models and triggering any post-merge actions.

The strategy for a merge is configurable, but currently
`FromSourceBranch` is the only available strategy. As a follow-up, we
will add a `FromTrainRef` strategy, which merges from a fully prepared
commit SHA associated with a successful pipeline on a merge train.