Fix storing messages for summarizing reviews
We fixed this before by setting `skip_cache: true` by default in
`ExecuteMethodService`. However, `SummarizeSubmittedReviewService` was
not going through the `ExecuteMethodService`.

As we only want to store messages from the `chat` action, and to fix
this for the future, the logic is now reversed and it's required to set
`cache_response: true` explicitly.