Fix truncate_legacy_tables for detached partitions
For some yet-to-be determined reason the FK constraints for these
"detached" partition tables are now stored in the
`gitlab_partitions_dynamic` schema. They used to be stored in the
`public` schema.

This affected the `truncate_legacy_tables` rake task as tables such as
`alert_management_alerts` are not truncated correctly.
See
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/1907.

As we have a mixture where some of these FKs are in
`gitlab_partitions_dynamic`, and some are in `public`, we need to query
both to get the foreign keys correctly.

Update specs with this scenario.

Changelog: fixed