Reduce inotify watches when cache_classes is enabled
This only affects development environments. Previously when Puma
started, 12 separate threads of `rb-fsevent` from the `listen` gem
were running to monitor file changes even when the `cache_classes`
setting were enabled. This occurred for two reasons:

1. The `factory_bot_rails` gem spins up file watchers, and there is no
way to disable this behavior. This appears to be responsible for 4
different threads.

2. The lookbook gem checks the `reload_classes_only_on_change` setting
to determine whether to spin up file watchers. While Rails ignores
this setting if `cache_classes` is enabled, the lookbook gem does not.
This appears to be responsible for 8 different threads.

Also disable these file watchers in Rake tasks or in the Rails console
to improve performance and to reduce the number of inotify watches
needed to run GitLab.

Changelog: changed