Refactor ci_config claims to use ClaimMapper
We plan to add additional ci_config claims for the `ExternalProject` and
`AutoDevops` configs. As preparation for this change, we are adding a
`ClaimMapper` class which can be used to produce the appropriate claim
value based on what type of config source has been provided. This lets
us avoid having to write a large case statement on `config.source`. This
change was not added to `ProjectConfig` itself because `ProjectConfig`
is used to _create_ Ci::Pipelines, and therefore it does not have access
to the pipeline model. Adding a mapper class lets us avoid ballooning
the responsibility of `ProjectConfig`.