Reword, move variable modal AWS/OIDC guidance
When modal shows a variable that's one of the AWS_TOKEN_CONSTANTS,
an alert should be shown with guidance on how to configure OIDC to get
these credentials instead of using variables.
This MR moves the alert to the top of the modal and updates its message
to be more generic with a link to the OIDC documentation.

Changelog: changed