Add `presubmissionCheckPassed` to `ExplainVulnerabilityPromptType`
The current behaviour is that if we detect a possible secret, we mark
the code as ineligible and prevent the generation of a prompt
containing it.

We want to give the user the power to choose to send it anyways.

So the secret detection check should no longer have any role in the
generation of the prompt.

What that leaves us with is only the `include_source_code` parameter
which can be `true`, `false` or `nil`.

If it's `true`, then we should always send a prompt including source
code as long as the file exists.

If it's `false`, then regardless of the presence of eligible code, we
want to send the codeless version of the prompt to the AI.

Finally, if it's `nil`, then we want to maintain the existing default behaviour
of explain this vulnerability, in which it will generate a prompt
including the source code if these two conditions are met:

1. there is eligible code available
2. no secrets have been detected

resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/417078

Changelog: changed
EE: true