Add new default places component
This is the first commit of a refacting of how the global search modal's
initial view is implemented.

This component is extracted from the `global_search_default_items.vue`
component, and represents the "Places" group of items shown in the
global search modal's initial view.

The idea is to extract each group displayed in the initial view to its
own component. Each component will:

- be responsible for getting its own data,
- render a GlDisclosureDropdownGroup as its root vnode,
- transparently pass all attrs to it (e.g., `bordered`),
- not render anything if they have no data,
- emit a `nothing-to-render` event if they have nothing to render.

A later commit will explain the rationale for these details.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/409522.