Add new default issuables component
This is the second commit of a refacting of how the global search
modal's initial view is implemented.

This component is extracted from the `global_search_default_items.vue`
component, and represents the context-dependent group of links to issues
and merge requests shown in the global search modal's initial view.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/409522.