Avoid caching patched node modules
If we cache patched node modules, whenever we change the patch,
it'll not apply to the patched modules. Patching is very fast so
this should not be an issue.

Using `after_script` is fine because we run it before uploading cache:
https://docs.gitlab.com/ee/ci/caching/#how-archiving-and-extracting-works

We also don't need to touch apt sources because it's left out from
d2cc841c55d65bc8134bfb3a467e66c36ac32b0a which was long ago. It's
needed when we're running `install_api_client_dependencies_with_apt`
and we're not doing that anymore.