Immediately render the Value Stream Dashboard data as it is loaded
Instead of waiting for all the data to finish loading before
rendering it in the table, insert the table data as soon as the
correlating request finishes.

Also only request metrics that were not excluded from the table.
Previously all metrics were requested even when excluded.

Changelog: performance
EE: true