Remove admin check when creating/updating subscription add-on purchases
Prior to this change, creating or updating a subscription add-on
purchase required the current user to be an admin. This check is being
done twice because it's also done as part of the API endpoint.

An incoming change for SM will not have a current user context when
creating/updating subscription add-on purchases. Therefore, it makes
sense to remove the admin check from those services and rely on the one
in the API endpoint.

Changelog: changed
EE: true