Prepare add-on purchase services to work without a namespace
Prior to this change, the services for
`GitlabSubscriptions::AddOnPurchases` relied on a namespace. With the
preparation to support code suggestions on SM, the services will be
reused. This means that the namespace requirement is only needed if the
code is running on GitLab.com. The validation for a namespace to be
present was already changed prior to this change.

This change modifies the services for
`GitlabSubscriptions::AddOnPurchases` to only check for a root namespace
if it's GitLab.com.

Additionally, this change also changes the services response text to be
more general and not API endpoint specific.

Changelog: changed
EE: true